// src/app/interface/iniciacion.interface.ts
// Generated by https://quicktype.io

export interface ResultIniciaciones {
  ok:         boolean;
  msg:        string;
  iniciacion: Iniciacion[];
}
export interface ResultIniciacion {
  ok:         boolean;
  msg:        string;
  iniciacion: Iniciacion;
}
export interface Iniciacion {
  id:     number;
  nombre: string;
  logo:   string;
  estado: number;
}

export interface IniciacionForm{
  nombre?:string;
  logo?:string;
}