// src/app/interface/tipo.iniciacion.interface.ts
// Generated by https://quicktype.io

// Generated by https://quicktype.io

export interface ResultTipoIniciaciones {
  ok:             boolean;
  msg:            string;
  tipoiniciacion: Tipoiniciacion[];
}
export interface ResultTipoIniciacion {
  ok:             boolean;
  msg:            string;
  tipoiniciacion: Tipoiniciacion;
}

export interface Tipoiniciacion {
  id:            number;
  descripcion:   string;
  id_iniciacion: number;
  Iniciacion:    Iniciacion;
}

export interface Iniciacion {
  id:     number;
  nombre: string;
  logo:   string;
  estado: number;
}